chore : "Review 클래스 변수 자료형 수정"

+++ src/main/java/yu/cse/odiga/store/domain/Review.java
@@ -20,9 +20,9 @@ public class Review {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
     private String content;
-    private double rating;
+    private int rating;
     private String imageUrl;
     private LocalDateTime createDate;
